Remove extraneous HTML from approval field
For some approvals the HPR site displays a badge that can be hovered to
see additional information about the approval. This leads to some stray
characters and carriage returns in our approval names.

This commit makes sure we strip away the badge (and any other HTML
elements) from the approval element before it's returned from our
 #approval method.