Fix keyword parameter usage
We have keyword param usage that is inconsistent between the
documentation and HPR.scraper() implementation as well as usage of
Fetcher#initialize() that is inconsistent with its implementation. This
starts failing under Ruby 3.

This commit adds a keyword parameter to HPR.scraper() definition, and
removes it from Fetcher.initialize().

The reason we now only use keyword param for HPR.scraper() is:

1) Add clarity for the end users of the library
2) Less verbosity when working with the internal classes.